+++ mitra_models/src/posts/queries.rs
@@ -839,7 +839,17 @@ pub async fn get_home_timeline(
                 UNION ALL
                 -- posts where user is mentioned
                 SELECT 1 FROM post_mention
-                WHERE post_id = post.id AND profile_id = $current_user_id
+                WHERE
+                    post_id = post.id
+                    AND profile_id = $current_user_id
+                    -- exclude mentions from muted conversations
+                    AND NOT EXISTS (
+                        SELECT 1 FROM conversation_tracking
+                        WHERE
+                            conversation_tracking.conversation_id = post.conversation_id
+                            AND account_id = $current_user_id
+                            AND tracking_status = {tracking_status_mute}
+                    )
                 UNION ALL
                 -- posts from followed conversations
                 SELECT 1 FROM conversation_tracking
@@ -860,6 +870,7 @@ pub async fn get_home_timeline(
         relationship_subscription=i16::from(RelationshipType::Subscription),
         relationship_hide_reposts=i16::from(RelationshipType::HideReposts),
         relationship_hide_replies=i16::from(RelationshipType::HideReplies),
+        tracking_status_mute=i16::from(TrackingStatus::Mute),
         tracking_status_follow=i16::from(TrackingStatus::Follow),
         mute_filter=build_mute_filter(),
         visibility_filter=build_visibility_filter(),
@@ -2100,6 +2111,10 @@ mod tests {
     use crate::{
         accounts::test_utils::create_test_user,
         activitypub::constants::AP_PUBLIC,
+        conversations::{
+            queries::set_conversation_tracking_status,
+            types::TrackingStatus,
+        },
         custom_feeds::queries::{
             add_custom_feed_sources,
             create_custom_feed,
@@ -2419,6 +2434,86 @@ mod tests {
         assert_eq!(timeline.iter().any(|post| post.id == post_13.id), false);
         assert_eq!(timeline.iter().any(|post| post.id == post_14.id), false);
         assert_eq!(timeline.len(), 8);
+    }
+
+    #[tokio::test]
+    #[serial]
+    async fn test_home_timeline_muted_conversation() {
+        let db_client = &mut create_test_database().await;
+        let current_user = create_test_user(db_client, "test").await;
+        let user_1 = create_test_user(db_client, "user1").await;
+        let user_2 = create_test_user(db_client, "user2").await;
+        follow(db_client, current_user.id, user_1.id).await.unwrap();
+        // Post from followed user
+        let post_data = PostCreateData {
+            content: "test 1".to_string(),
+            ..Default::default()
+        };
+        let post = create_post(db_client, user_1.id, post_data).await.unwrap();
+        // Reply from current user
+        let reply_data_1 = PostCreateData {
+            context: PostContext::reply_to(&post),
+            content: "reply".to_string(),
+            mentions: vec![user_1.id],
+            ..Default::default()
+        };
+        let reply_1 = create_post(db_client, current_user.id, reply_data_1).await.unwrap();
+        // Reply from followed user
+        let reply_data_2 = PostCreateData {
+            context: PostContext::reply_to(&reply_1),
+            content: "reply".to_string(),
+            mentions: vec![current_user.id],
+            ..Default::default()
+        };
+        let reply_2 = create_post(db_client, user_1.id, reply_data_2).await.unwrap();
+        // Reply from 3rd user
+        let reply_data_3 = PostCreateData {
+            context: PostContext::reply_to(&reply_2),
+            content: "reply".to_string(),
+            mentions: vec![current_user.id, user_1.id],
+            ..Default::default()
+        };
+        let reply_3 = create_post(db_client, user_2.id, reply_data_3).await.unwrap();
+        // Reply from followed user
+        let reply_data_4 = PostCreateData {
+            context: PostContext::reply_to(&reply_3),
+            content: "reply".to_string(),
+            mentions: vec![current_user.id, user_2.id],
+            ..Default::default()
+        };
+        let reply_4 = create_post(db_client, user_1.id, reply_data_4).await.unwrap();
+
+        let timeline = get_home_timeline(
+            db_client,
+            current_user.id,
+            None,
+            20,
+        ).await.unwrap();
+        assert_eq!(timeline.len(), 5);
+        assert_eq!(timeline[0].id, reply_4.id);
+        assert_eq!(timeline[1].id, reply_3.id);
+        assert_eq!(timeline[2].id, reply_2.id);
+        assert_eq!(timeline[3].id, reply_1.id);
+        assert_eq!(timeline[4].id, post.id);
+
+        set_conversation_tracking_status(
+            db_client,
+            post.expect_conversation().id,
+            current_user.id,
+            Some(TrackingStatus::Mute),
+        ).await.unwrap();
+        let timeline = get_home_timeline(
+            db_client,
+            current_user.id,
+            None,
+            20,
+        ).await.unwrap();
+        assert_eq!(timeline.len(), 4);
+        // Reply from 3rd user is not shown
+        assert_eq!(timeline[0].id, reply_4.id);
+        assert_eq!(timeline[1].id, reply_2.id);
+        assert_eq!(timeline[2].id, reply_1.id);
+        assert_eq!(timeline[3].id, post.id);
     }
 
     #[tokio::test]