+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `/api/v1/pleroma/notifications/read` API endpoint.
 - Added `pleroma.is_seen` field to `Notification` API entity.
+- Added ability to mute a conversation.
 
 ### Changed